Add support for using GitHub token from .gitconfig
This utility makes API queries, and depending on your source IP or your
use of the API, you may be throttled and limited in the numer of API
queries you can make. The solution is to add a token for GitHub in your
`.gitconfig` file as [documented here](https://github.blog/2008-10-11-local-github-config/).

git clone-fork will use this token when talking to GitHub.